% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inference_boot}
\alias{inference_boot}
\title{Compute confidence intervals from results of bootstrapping.}
\usage{
inference_boot(res_boot)
}
\arguments{
\item{res_boot}{Result from a call to \code{boot::boot}. A named list.}
}
\value{
Tibble with point estimate and confidence interval.
}
\description{
Compute confidence intervals from results of bootstrapping.
}
