% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{ipw}
\alias{ipw}
\title{Implementations of IPW estimators.}
\usage{
ipw(
  dat,
  idxs,
  baseline,
  time_varying,
  exposures,
  outcome,
  censoring,
  forms,
  form_cens,
  shift_functions,
  survey_weights,
  SL_library,
  cross_val,
  method
)
}
\arguments{
\item{dat}{Dataset with confounders, exposures, and outcome. A dataframe.}

\item{idxs}{Indexes for bootstrapping. A vector of integers.}

\item{baseline}{Baseline confounders. A vector of strings.}

\item{time_varying}{Time-varying confounders. A vector of vectors of strings.}

\item{exposures}{Exposure variables. A vector of vectors of strings.}

\item{outcome}{Outcome variable. A string.}

\item{censoring}{Censoring variable. A vector of strings.}

\item{forms}{Models formulas. A vector of strings.}

\item{form_cens}{TBD}

\item{shift_functions}{Functions to shift the exposures. A vector of functions.}

\item{survey_weights}{Optional vector of survey/balancing weights. A vector.}

\item{SL_library}{Vector of learners or function. A vector of strings or function.}

\item{cross_val}{TBD.}

\item{method}{Which method to use for estimating the weights. A string.}
}
\value{
The counterfactual mean of the outcome.
}
\description{
Implementations of IPW estimators.
}
