% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{ipw_bin_fix}
\alias{ipw_bin_fix}
\title{Implementation of a specific IPW estimator
for a single binary exposure and time point.}
\usage{
ipw_bin_fix(
  dat,
  idxs,
  baseline,
  exposure,
  outcome,
  form,
  form_cens,
  survey_weights,
  SL_library,
  cross_val
)
}
\arguments{
\item{dat}{Dataset with confounders, exposure, and outcome. A dataframe.}

\item{idxs}{Indexes for bootstrapping. A vector of integers.}

\item{baseline}{Baseline confounders. A vector of strings.}

\item{exposure}{Exposure variable. A string.}

\item{outcome}{Outcome variable. A string.}

\item{form}{Model formula. A string.}

\item{form_cens}{TBD}

\item{survey_weights}{Optional vector of survey/balancing weights. A vector.}

\item{SL_library}{Vector of learners or function. A vector of strings or function.}

\item{cross_val}{TBD.}
}
\value{
The counterfactual means of the outcome under no intervention
(natural course) and the intervention of interest as
specified by the shift function.
}
\description{
Implementation of a specific IPW estimator
for a single binary exposure and time point.
}
