% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp.R
\name{parametric_g_computation}
\alias{parametric_g_computation}
\title{Implementation of the parametric g-formula.}
\usage{
parametric_g_computation(
  dat,
  idxs,
  baseline,
  time_varying,
  exposures,
  outcome,
  censoring,
  forms,
  shift_functions,
  survey_weights,
  SL_library,
  cross_val
)
}
\arguments{
\item{dat}{Dataset with confounders, exposures, and outcome. A dataframe.}

\item{idxs}{Indexes for bootstrapping. A vector of integers.}

\item{baseline}{Baseline confounders. A vector of strings.}

\item{time_varying}{Time-varying confounders. A vector of vectors of strings.}

\item{exposures}{Exposure variables. A vector of vectors of strings.}

\item{outcome}{Outcome variable. A string.}

\item{censoring}{Censoring variable. A vector of strings.}

\item{forms}{Models formulas. A vector of strings.}

\item{shift_functions}{Functions to shift the exposures. A vector of functions.}

\item{survey_weights}{Optional vector of survey/balancing weights. A vector.}

\item{SL_library}{Vector of learners or function. A vector of strings or function.}

\item{cross_val}{TBD.}
}
\value{
The counterfactual mean of the outcome.
}
\description{
Implementation of the parametric g-formula.
}
